Does not actually fix a problem.
When I run `koch tools` with gcc/4.9.2 on centos6 (for distributable
binaries), I get the "iso scanf" problem for linenoise.c. I
want `-D_GNU_SOURCE`, but this is not enough. So I "koch tools" does
not finish. I guess it is not important. But if you ever fix this,
you can update this branch.